# server/docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "使用gin+vue进行极速开发的全栈开发基础平台",
    "title": "Gin-Vue-Admin Swagger API接口文档",
    "contact": {},
    "version": "v2.5.9"
  },
  "paths": {
...
    "/authority/copyAuthority": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Authority"],
        "summary": "拷贝角色",
        "parameters": [
          {
            "description": "旧角色id, 新权限id, 新权限名, 新父角色id",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/response.SysAuthorityCopyResponse"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "拷贝角色,返回包括系统角色详情",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/response.SysAuthorityResponse"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/authority/createAuthority": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Authority"],
        "summary": "创建角色",
        "parameters": [
          {
            "description": "权限id, 权限名, 父角色id",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/system.SysAuthority"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "创建角色,返回包括系统角色详情",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/response.SysAuthorityResponse"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/authority/deleteAuthority": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Authority"],
        "summary": "删除角色",
        "parameters": [
          {
            "description": "删除角色",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/system.SysAuthority"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除角色",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/authority/getAuthorityList": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Authority"],
        "summary": "分页获取角色列表",
        "parameters": [
          {
            "description": "页码, 每页大小",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/request.PageInfo"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "分页获取角色列表,返回包括列表,总数,页码,每页数量",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/response.PageResult"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/authority/setDataAuthority": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Authority"],
        "summary": "设置角色资源权限",
        "parameters": [
          {
            "description": "设置角色资源权限",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/system.SysAuthority"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "设置角色资源权限",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
    "/authority/updateAuthority": {
      "post": {
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "tags": ["Authority"],
        "summary": "更新角色信息",
        "parameters": [
          {
            "description": "权限id, 权限名, 父角色id",
            "name": "data",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/system.SysAuthority"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "更新角色信息,返回包括系统角色详情",
            "schema": {
              "allOf": [
                {
                  "$ref": "#/definitions/response.Response"
                },
                {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/definitions/response.SysAuthorityResponse"
                    },
                    "msg": {
                      "type": "string"
                    }
                  }
                }
              ]
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "request.PageInfo": {
      "type": "object",
      "properties": {
        "keyword": {
          "description": "关键字",
          "type": "string"
        },
        "page": {
          "description": "页码",
          "type": "integer"
        },
        "pageSize": {
          "description": "每页大小",
          "type": "integer"
        }
      }
    },
...
    "response.PageResult": {
      "type": "object",
      "properties": {
        "list": {},
        "page": {
          "type": "integer"
        },
        "pageSize": {
          "type": "integer"
        },
        "total": {
          "type": "integer"
        }
      }
    },
...
    "response.Response": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer"
        },
        "data": {},
        "msg": {
          "type": "string"
        }
      }
    },
...
    "response.SysAuthorityCopyResponse": {
      "type": "object",
      "properties": {
        "authority": {
          "$ref": "#/definitions/system.SysAuthority"
        },
        "oldAuthorityId": {
          "description": "旧角色ID",
          "type": "integer"
        }
      }
    },
    "response.SysAuthorityResponse": {
      "type": "object",
      "properties": {
        "authority": {
          "$ref": "#/definitions/system.SysAuthority"
        }
      }
    },
...
    "system.Meta": {
      "type": "object",
      "properties": {
        "activeName": {
          "type": "string"
        },
        "closeTab": {
          "description": "自动关闭tab",
          "type": "boolean"
        },
        "defaultMenu": {
          "description": "是否是基础路由（开发中）",
          "type": "boolean"
        },
        "icon": {
          "description": "菜单图标",
          "type": "string"
        },
        "keepAlive": {
          "description": "是否缓存",
          "type": "boolean"
        },
        "title": {
          "description": "菜单名",
          "type": "string"
        }
      }
    },
...
    "system.SysAuthority": {
      "type": "object",
      "properties": {
        "authorityId": {
          "description": "角色ID",
          "type": "integer"
        },
        "authorityName": {
          "description": "角色名",
          "type": "string"
        },
        "children": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/system.SysAuthority"
          }
        },
        "createdAt": {
          "description": "创建时间",
          "type": "string"
        },
        "dataAuthorityId": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/system.SysAuthority"
          }
        },
        "defaultRouter": {
          "description": "默认菜单(默认dashboard)",
          "type": "string"
        },
        "deletedAt": {
          "type": "string"
        },
        "menus": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/system.SysBaseMenu"
          }
        },
        "parentId": {
          "description": "父角色ID",
          "type": "integer"
        },
        "updatedAt": {
          "description": "更新时间",
          "type": "string"
        }
      }
    },
...
    "system.SysBaseMenu": {
      "type": "object",
      "properties": {
        "ID": {
          "description": "主键ID",
          "type": "integer"
        },
        "authoritys": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/system.SysAuthority"
          }
        },
        "children": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/system.SysBaseMenu"
          }
        },
        "component": {
          "description": "对应前端文件路径",
          "type": "string"
        },
        "createdAt": {
          "description": "创建时间",
          "type": "string"
        },
        "hidden": {
          "description": "是否在列表隐藏",
          "type": "boolean"
        },
        "menuBtn": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/system.SysBaseMenuBtn"
          }
        },
        "meta": {
          "description": "附加属性",
          "allOf": [
            {
              "$ref": "#/definitions/system.Meta"
            }
          ]
        },
        "name": {
          "description": "路由name",
          "type": "string"
        },
        "parameters": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/system.SysBaseMenuParameter"
          }
        },
        "parentId": {
          "description": "父菜单ID",
          "type": "string"
        },
        "path": {
          "description": "路由path",
          "type": "string"
        },
        "sort": {
          "description": "排序标记",
          "type": "integer"
        },
        "updatedAt": {
          "description": "更新时间",
          "type": "string"
        }
      }
    },
    "system.SysBaseMenuBtn": {
      "type": "object",
      "properties": {
        "ID": {
          "description": "主键ID",
          "type": "integer"
        },
        "createdAt": {
          "description": "创建时间",
          "type": "string"
        },
        "desc": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "sysBaseMenuID": {
          "type": "integer"
        },
        "updatedAt": {
          "description": "更新时间",
          "type": "string"
        }
      }
    },
    "system.SysBaseMenuParameter": {
      "type": "object",
      "properties": {
        "ID": {
          "description": "主键ID",
          "type": "integer"
        },
        "createdAt": {
          "description": "创建时间",
          "type": "string"
        },
        "key": {
          "description": "地址栏携带参数的key",
          "type": "string"
        },
        "sysBaseMenuID": {
          "type": "integer"
        },
        "type": {
          "description": "地址栏携带参数为params还是query",
          "type": "string"
        },
        "updatedAt": {
          "description": "更新时间",
          "type": "string"
        },
        "value": {
          "description": "地址栏携带参数的值",
          "type": "string"
        }
      }
    },
...
  },
  "securityDefinitions": {
    "ApiKeyAuth": {
      "type": "apiKey",
      "name": "x-token",
      "in": "header"
    }
  }
}